requirements.txt
flask
openai
sentence-transformers
langchain
pandas
faiss-cpu
numpy
scipy
jsonschema
PyPDF2
huggingface_hub
python-dotenv
langchain_community
langchain_openai
flask_cors
colorlog
werkzeug
anthropic